Generate model files as part of build process
Replace the manual model generation step with a gradle build step which
uses openapi codegen to generate model files.

To generate a new version of the api models, download the most recent
swagger spec from https://github.com/typesense/typesense-api-spec and
place in the root of this repo.

Moved from jaxrs-cxf-client to java/jersey2. CXF was an unused
dependency, so choosing a generator library which includes the jackson
mapper used in the api call class